+++ CHANGELOG.md
@@ -35,6 +35,11 @@ it is time to give credit to original [author](https://github.com/auTOMATIC1111)
     Simply enable checkbox in advanced menu and set desired resolution  
     Additional settings are available in *settings -> inference settings -> hidiffusion*  
     And can also be set and used via *xyz grid*  
+  - [Perturbed-Attention Guidance](https://github.com/KU-CVLAB/Perturbed-Attention-Guidance)  
+    PAG enhances sample quality by utilizing self-attention in formation of latent in addition to standard CFG scale  
+    Simply set *advanced -> attention guidance* and *advanced -> adaptive scaling*  
+    Additional options are available in *settings -> inference settings -> pag*  
+    *Note*: PAG has replaced SAG as attention guidance method in SD.Next  
   - **IP Adapter Masking**:  
     Powerful method of using masking with ip-adapters  
     When combined with multiple ip-adapters, it allows for different inputs guidance for each segment of the input image